Moving filterEvents from ui.js to  api.js
